+++ src/main/java/pl/asie/computronics/block/BlockDigitalReceiverBox.java
@@ -1,6 +1,8 @@
 package pl.asie.computronics.block;
 
 import cpw.mods.fml.client.registry.RenderingRegistry;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import li.cil.oc.api.network.Environment;
 import mods.railcraft.api.core.IPostConnection;
 import mods.railcraft.client.util.textures.TextureAtlasSheet;
@@ -30,14 +32,14 @@ import java.util.List;
 /**
  * @author CovertJaguar, Vexatos
  */
-public class BlockDigitalReceiver extends BlockContainer
+public class BlockDigitalReceiverBox extends BlockContainer
 	implements IPostConnection {
 
 	public static IIcon[] texturesBox;
 	public static IIcon texturesBoxTop;
 	private final int renderType;
 
-	public BlockDigitalReceiver() {
+	public BlockDigitalReceiverBox() {
 		super(new MaterialStructure());
 		this.renderType = RenderingRegistry.getNextAvailableRenderId();
 		setBlockName("computronics.signalBox");
@@ -46,20 +48,26 @@ public class BlockDigitalReceiver extends BlockContainer
 		setCreativeTab(Computronics.tab);
 	}
 
+	@Override
 	@SuppressWarnings("unchecked")
 	public void getSubBlocks(Item item, CreativeTabs tab, List list) {
 		list.add(new ItemStack(item, 1, 0));
 	}
 
+	@Override
 	public int damageDropped(int meta) {
 		return meta;
 	}
 
+	@Override
+	@SideOnly(Side.CLIENT)
 	public void registerBlockIcons(IIconRegister iconRegister) {
 		texturesBox = TextureAtlasSheet.unstitchIcons(iconRegister, "railcraft:signal.box", 6);
 		texturesBoxTop = iconRegister.registerIcon("computronics:signal_box_receiver");
 	}
 
+	@Override
+	@SideOnly(Side.CLIENT)
 	public IIcon getIcon(int side, int meta) {
 		switch(side){
 			case 0:
@@ -71,6 +79,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		}
 	}
 
+	@Override
 	public boolean onBlockActivated(World world, int i, int j, int k, EntityPlayer player, int side, float u1, float u2, float u3) {
 		ItemStack current = player.getCurrentEquippedItem();
 		if((current != null) &&
@@ -82,12 +91,14 @@ public class BlockDigitalReceiver extends BlockContainer
 			&& ((TileDigitalReceiverBox) tile).blockActivated(side, player);
 	}
 
+	@Override
 	public boolean rotateBlock(World world, int x, int y, int z, ForgeDirection axis) {
 		TileEntity tile = world.getTileEntity(x, y, z);
 		return (tile instanceof TileDigitalReceiverBox)
 			&& ((TileDigitalReceiverBox) tile).rotateBlock(axis);
 	}
 
+	@Override
 	public ForgeDirection[] getValidRotations(World world, int x, int y, int z) {
 		TileEntity tile = world.getTileEntity(x, y, z);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -96,6 +107,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		return super.getValidRotations(world, x, y, z);
 	}
 
+	@Override
 	public void onPostBlockPlaced(World world, int i, int j, int k, int meta) {
 		super.onPostBlockPlaced(world, i, j, k, meta);
 		TileEntity tile = world.getTileEntity(i, j, k);
@@ -104,6 +116,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		}
 	}
 
+	@Override
 	public void onBlockPlacedBy(World world, int i, int j, int k, EntityLivingBase entityliving, ItemStack stack) {
 		TileEntity tile = world.getTileEntity(i, j, k);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -111,6 +124,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		}
 	}
 
+	@Override
 	public void onNeighborBlockChange(World world, int x, int y, int z, Block block) {
 		try {
 			TileEntity tile = world.getTileEntity(x, y, z);
@@ -130,6 +144,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		}
 	}
 
+	@Override
 	public void breakBlock(World world, int i, int j, int k, Block block, int meta) {
 		TileEntity tile = world.getTileEntity(i, j, k);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -138,6 +153,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		super.breakBlock(world, i, j, k, block, meta);
 	}
 
+	@Override
 	public void setBlockBoundsBasedOnState(IBlockAccess world, int i, int j, int k) {
 		TileEntity tile = world.getTileEntity(i, j, k);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -147,6 +163,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		}
 	}
 
+	@Override
 	public AxisAlignedBB getCollisionBoundingBoxFromPool(World world, int i, int j, int k) {
 		TileEntity tile = world.getTileEntity(i, j, k);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -156,6 +173,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		return super.getCollisionBoundingBoxFromPool(world, i, j, k);
 	}
 
+	@Override
 	public AxisAlignedBB getSelectedBoundingBoxFromPool(World world, int x, int y, int z) {
 		TileEntity tile = world.getTileEntity(x, y, z);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -164,6 +182,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		return AxisAlignedBB.getBoundingBox(x + this.minX, y + this.minY, z + this.minZ, x + this.maxX, y + this.maxY, z + this.maxZ);
 	}
 
+	@Override
 	public int getLightValue(IBlockAccess world, int x, int y, int z) {
 		if(y < 0) {
 			return 0;
@@ -175,6 +194,7 @@ public class BlockDigitalReceiver extends BlockContainer
 		return 0;
 	}
 
+	@Override
 	public float getBlockHardness(World world, int x, int y, int z) {
 		TileEntity tile = world.getTileEntity(x, y, z);
 		if((tile instanceof TileDigitalReceiverBox)) {
@@ -183,60 +203,74 @@ public class BlockDigitalReceiver extends BlockContainer
 		return 3.0F;
 	}
 
+	@Override
 	public boolean isSideSolid(IBlockAccess world, int i, int j, int k, ForgeDirection side) {
 		TileEntity tile = world.getTileEntity(i, j, k);
 		return (tile instanceof TileDigitalReceiverBox)
 			&& ((TileDigitalReceiverBox) tile).isSideSolid(world, i, j, k, side);
 	}
 
+	@Override
 	public boolean isOpaqueCube() {
 		return false;
 	}
 
+	@Override
 	public boolean renderAsNormalBlock() {
 		return false;
 	}
 
+	@Override
 	public int getRenderType() {
 		return this.renderType;
 	}
 
+	@Override
 	public TileEntity createNewTileEntity(World var1, int meta) {
-		return new TileDigitalReceiverBox("digital_receiver_box");
+		return new TileDigitalReceiverBox();
 	}
 
+	@Override
 	public boolean canProvidePower() {
 		return false;
 	}
 
+	@Override
 	public boolean canConnectRedstone(IBlockAccess world, int i, int j, int k, int dir) {
 		return false;
 	}
 
+	@Override
 	public boolean isBlockNormalCube() {
 		return false;
 	}
 
+	@Override
 	public boolean shouldSideBeRendered(IBlockAccess world, int x, int y, int z, int side) {
 		return true;
 	}
 
+	@Override
 	public int isProvidingWeakPower(IBlockAccess world, int i, int j, int k, int side) {
 		return 0;
 	}
 
+	@Override
 	public boolean hasTileEntity(int metadata) {
 		return true;
 	}
 
+	@Override
 	public boolean canBeReplacedByLeaves(IBlockAccess world, int x, int y, int z) {
 		return false;
 	}
 
+	@Override
 	public boolean canCreatureSpawn(EnumCreatureType type, IBlockAccess world, int x, int y, int z) {
 		return false;
 	}
 
+	@Override
 	public IPostConnection.ConnectStyle connectsToPost(IBlockAccess world, int x, int y, int z, ForgeDirection side) {
 		TileEntity t = world.getTileEntity(x, y, z);
 		if((t instanceof ISignalTile)) {