+++ src/main/java/pl/asie/computronics/cc/IComputronicsPeripheral.java
@@ -1,0 +1,12 @@
+package pl.asie.computronics.cc;
+
+import dan200.computercraft.api.peripheral.IPeripheral;
+
+/**
+ * Only used to check whether a peripheral belongs to Computronics
+ * @author Vexatos
+ */
+public interface IComputronicsPeripheral extends IPeripheral {
+
+}
+