+++ src/main/java/pl/asie/computronics/cc/ISidedPeripheral.java
@@ -1,0 +1,10 @@
+package pl.asie.computronics.cc;
+
+/**
+ * @author Vexatos
+ */
+public interface ISidedPeripheral {
+
+	public boolean canConnectPeripheralOnSide(int side);
+
+}