+++ src/main/java/pl/asie/computronics/reference/Names.java
@@ -9,6 +9,8 @@ public class Names {
 	public static final String FSP_SteamTransporter = "steam_transporter";
 	public static final String FZ_ChargeConductor = "charge_conductor";
 	public static final String MFR_DSU = "deep_storage_unit";
+	public static final String Railcraft_DigitalDetector = "digital_detector";
+	public static final String Railcraft_DigitalReceiverBox = "digital_receiver_box";
 	public static final String Railcraft_ElectricGrid = "electric_tile";
 	public static final String Railcraft_LimiterTrack = "limiter_track";
 	public static final String Railcraft_LocomotiveTrack = "locomotive_track";